cartoonify/requirements.txt
contourpy==1.3.0
cycler==0.12.1
easygui==0.98.3
fonttools==4.53.1
imageio==2.35.1
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.1.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
